+++ lib/segment/src/id_tracker/mutable_id_tracker/mod.rs
@@ -216,13 +216,13 @@ impl IdTrackerRead for MutableIdTracker {
 
     fn iter_internal_versions(
         &self,
-    ) -> Box<dyn Iterator<Item = (PointOffsetType, SeqNumberType)> + '_> {
-        Box::new(
+    ) -> OperationResult<Box<dyn Iterator<Item = (PointOffsetType, SeqNumberType)> + '_>> {
+        Ok(Box::new(
             self.internal_to_version
                 .iter()
                 .enumerate()
                 .map(|(i, version)| (i as PointOffsetType, *version)),
-        )
+        ))
     }
 
     fn name(&self) -> &'static str {
